refactor(organization): add dialog components for logo management and enhance organization form
