--- features/organization/insex.ts
@@ -1,3 +1,0 @@
-export * from './organization-card';
-export * from './organization-form';
-export * from './organization-selector';
